requirements.txt
discord.py
argparse